Successfully render all information on all routes with date functions moved to helpers file
