+++ src/Views/Home.vue
@@ -1,11 +1,45 @@
 <template>
   <div>
     <h1>Photo of the Day</h1>
-    <MonthContainer />
+    <PhotoDetail v-bind:current="current" />
   </div>
 </template>
 
 <script>
-import MonthContainer from '../components/MonthContainer.vue';
-export default {};
+import PhotoDetail from '../components/PhotoDetail.vue';
+import { getTodaysPhoto } from '../utils/apiCalls/apiCalls';
+
+export default {
+  name: 'Home',
+  components: {
+    PhotoDetail
+  },
+  data() {
+    return {
+      current: {},
+      displayDate: ''
+    };
+  },
+  methods: {
+    async setTodayPhoto() {
+      try {
+        let photo = await getTodaysPhoto();
+        this.current = photo;
+      } catch (error) {
+        null;
+      }
+    }
+  },
+  mounted() {
+    this.setTodayPhoto();
+  }
+};
 </script>
+
+<style scoped>
+* {
+  box-sizing: border-box;
+  margin: 0;
+  padding: 0;
+}
+</style>
